Deduplicate style and preview path helpers
