+++ app/src/lib/build-styles.ts
@@ -10,6 +10,11 @@
 
 import * as fs from "node:fs/promises";
 import * as path from "node:path";
+import {
+  isWildcard,
+  splitVariantSegments,
+  toRegexFromWildcard,
+} from "./styles-shared.ts";
 import type {
   AtPropertyDef,
   ModuleJson,
@@ -507,47 +512,6 @@ interface GlobalIndex {
   variantToModule: Map<string, string>;
   variantWildcards: Array<{ name: string; module: string; re: RegExp }>;
   atPropToModule: Map<string, string>;
-}
-
-function isWildcard(name: string): boolean {
-  return name.includes("*");
-}
-
-function escapeRegexSpecialChars(input: string) {
-  return input.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-
-function toRegexFromWildcard(pattern: string) {
-  let regexBody = "";
-  for (let i = 0; i < pattern.length; i++) {
-    const char = pattern.charAt(i);
-    regexBody += char === "*" ? ".*" : escapeRegexSpecialChars(char);
-  }
-  return new RegExp(`^${regexBody}$`);
-}
-
-function splitVariantSegments(token: string) {
-  const segments: string[] = [];
-  let start = 0;
-  let bracketDepth = 0;
-  let parenDepth = 0;
-  for (let i = 0; i < token.length; i++) {
-    const char = token.charAt(i);
-    if (char === "[") {
-      bracketDepth++;
-    } else if (char === "]") {
-      bracketDepth = Math.max(0, bracketDepth - 1);
-    } else if (char === "(") {
-      parenDepth++;
-    } else if (char === ")") {
-      parenDepth = Math.max(0, parenDepth - 1);
-    } else if (char === ":" && bracketDepth === 0 && parenDepth === 0) {
-      segments.push(token.slice(start, i));
-      start = i + 1;
-    }
-  }
-  segments.push(token.slice(start));
-  return segments;
 }
 
 function buildGlobalIndex(modules: ModuleJson[]): GlobalIndex {
